+++ packages/argdown-node/CHANGELOG.md
@@ -1,7 +1,18 @@
 # @argdown/node
 
-## 1.0.0 (10-16-2018)
+## 1.1.0 (2019-02-26)
 
-* renamed and split of from argdown-cli
-* switched to Typescript
-* support for @argdown/core 1.0.0
+### Minor changes
+
+- updated packages
+- added "engine" and "nop" setting to DotToSvgExportPlugin to support Viz.js options (all Graphviz layout engines can now be used)
+
+### Bug fixes
+
+- fixes #104 Extension host stops in VSCODE
+
+## 1.0.0 (2018-10-16)
+
+- renamed and split of from argdown-cli
+- switched to Typescript
+- support for @argdown/core 1.0.0